[SNOW-726924] Add jvm.nonProxy.hosts Parameter and update JDBC to 3.13.23
This parameter allows users to bypass the proxy server for certain hosts
by setting the JVM argument -Dhttp.nonProxyHosts. For reference:
https://docs.oracle.com/javase/8/docs/technotes/guides/net/proxies.html
Older JDBC drivers ignored the http.nonProxyHosts parameter so the version
was upgraded from 3.13.14 to 3.13.23. The ingest-sdk does not honor
the parameter either as of v0.10.8. Therefore, using this parameter
or the JVM argument in versions 1.6.9 or lower of the connector will
cause a failure if you add *.snowflakecomputing.com. It's safe for
any other hosts.