[SNOW-726924] Add New jvm.nonProxy.hosts Parameter and update JDBC to 3.13.23
The jvm.nonProxy.hosts parameter can be added to the connector's configuration
file and is the equivalent to the JVM argument -Dhttp.nonProxyHosts. If a user
provides that parameter and the connector runs in a JVM where -Dhttp.nonProxyHosts
was set, then we will merge the two lists of hosts. In the future, we'll want another
method of adding proxy configurations without resorting to System.setProperty() calls